Provide a first version of an Instrument class
The Instrument class is an abstract class that can be inherited from to create various instruments, such as oscillators, samplers, and more. The whole proposed design is represented in the UML diagram provided in this commit. To use this class, simply inherit from it and define the renderNextAudioBlock() function to fill in the internal buffer.

The purpose of the Instrument class is to manage an internal set of parameters such as the frequency the instrument should play, and the volume at which this frequency should be played. The Instrument class automatically smooths out any parameter change unless instructed otherwise, in order to avoid undesirable audio artifacts.

One can inherit from either Instrument, to obtain an instrument with already three parameters inside (frequency to play, velocity, and volume), or directly from AbstractInstrument to obtain an empty instrument with no predefined parameter inside.