New config directory, and structure of Instrument.cpp
A new configuration directory was created in the repository, in order to provide a centralized and organized interface for custom macros. The implementation of the Instrument.cpp file was the occasion to start populating that directory with an AudioConfig.h file, which features only one macro for the moment for audio instruments.