Explicited a cast from double to uint32_t when updating the sampleRate
Note that if sampleRate*m_minSmoothingDuration is not an integer, then the static_cast will truncate the number of samples to use for parameter smoothing. To be more accurate, we would therefore need an extra sample to properly cover the whole smoothing duration, but this was not implemented since this truncation would rarely happen in practice (if m_minSmoothingDuration is equal to an integer when expressed in milliseconds and sampleRate is equal to an integer when expressed in kHz, which is usually the case, then the multiplication will result in an integer).