Adding a comment on the absence of integer overflow
Computing the remaining samples of a transient consists in subtracting the current position to the total number of samples to iterate over:

uint32_t remainingSamples = transientTracker.transientDurationInSamples - transientTracker.position

This type of operation can lead to overflow, but here in the code, the parameter is in a TRANSIENT state and no change request has been received. This means that the only method which has changed the parameter's internal components is the updateParametersAfterRendering() method, which should have pushed the parameter into a STEADY state if we had the inequality: position >= transientDurationInSamples. Since it did not, it means we have the opposite inequality: position < transientDurationInSamples, and therefore we will not have any overflow.