Finished implementing all the methods of the BaseInstrument class
Note that requesting an instantaneous parameter change will actually take effect immediately, and will not lead to posting any ChangeRequest. This is to ensure the end-user gets zero latency when requesting an instant change.

This implies that part of the code of the renderParametersForNextAudioBlock method will never been reached (when testing if the change request is smooth results in false). Nonetheless, if the "instantaneous change" policy eventually evolves due to concurrency issues, this also means the code will be ready for adaptation. One will simply have to change the requestParameterChange method into posting a ChangeRequest every time, no matter whether the change is be smooth or not, which will be a lot safer regarding concurrency.

A new parameter() method was added, in order to access a parameter during rendering without testing its state on each access.