Implemented the minimal smoothing feature
The ParameterRenderer now uses the minimalSmoothingEnabled's attibute of all parameters to configure its rendering when proceeding a ChangeRequest. This feature enables the ParameterRenderer to actually smooth out any change that was not required to be smooth by the end-user, but which actually needs to be smoothed out a minimum in order to be inaudible.

The following table explains the behavior of the ParameterRenderer in every possible case when running its renderParametersForNextAudioBlock() method.
Note that a ChangeRequest that is required to be smooth but should last 0 samples is treated an instantaneous change request. Also note that in case of an instantaneous ChangeRequest, the specified number of samples is ignored.

Legend:
The first column indicates whether the change request received should be smooth or not.
The second column indicates the number of samples during which the change request should be performed on according to the end-user.
The third column indicates whether the minimal smoothing feature is on for the Parameter requested to change.
The fourth column indicates the action that the ParameterRender will take.

                | numSamples=0  | Minimal Smoothing | perform a smooth change of length minSmoothingSamples
                |               |        -          | ignored
Smooth Change   |--------------------------------------------------------------------------------------------
                | numSamples>0  | Minimal Smoothing | perform a smooth change of length max(minSmoothingSamples, changeRequestDuration)
                |               |        -          | perform a smooth change of length changeRequestDuration
-------------------------------------------------------------------------------------------------------------
                | numSamples=0  | Minimal Smoothing | perform a smooth change of length minSmoothingSamples
                |               |        -          | ignored (precedence of shouldBeSmooth over numSamples)
Instant. Change |--------------------------------------------------------------------------------------------
                | numSamples>0  | Minimal Smoothing | perform a smooth change of length minSmoothingSamples
                |               |        -          | ignored (precedence of shouldBeSmooth over numSamples)

All the cases corresponding to "ignored" in the above table are taken into account in the requestParameterChange() method, in the form of an immediate change with zero latency.