Restructure the transient curve computation
The transient curve's increment is now computed in 3 steps, which are clearer to follow in the code.

The fmax method was inappropriately used and was changed for the std::max method instead, which is from the standard C++ algorithm header, and which can be specialized for comparing uint32_t.

The minimum level for multiplicative increments is also used for computing the transient's startValue, which was not the case before (the parameter's internal value was directly taken as a start value, despite the fact it could equal zero!).

Finally, the Instrument now uses the fast exponential approximation from the FastMath class to compute the multiplicative increment of a geometric transient curve.