Began reimplementing the library's core
This branch marks the renewal of the library's core. We are now heading to a fast, lock-free, and mutatable audio worfklow.

The library's header file has been reset to zero, and now only displays the tested features of the new core.

For the moment, only the Stream and WorkflowItem features have been tested. One other feature has been implemented (the Worker class), and will be tested later.