Added a hasInputs() method to the Worker class
This method will be used to determine if a Worker is a Generator (i.e. a leaf in the rendering tree) when exploring the Workflow.

Other modifications include cosmetic changes, and rewrite of the Worker class's constructor to perform a single allocation per bus and use the algorithm features of the standard library.