Added checks and comments in the Mixer's work method
Just like any Worker, the Mixer should check if its input and output stream pointers are nullptr before reading or writing data. Altough this verification is sometimes unnecessary (for example, when a Worker such as an Instrument is always connected to its environment on every port), it is mandatory for the Mixer, who could have some unplugged input entries depending on the number of instruments which are created and destroyed at runtime.