Extracted the voice assignment feature into a dedicated class
VoiceAssigner assigns voice numbers to workers.

VoiceManager will manage the voices themselves, and proceed to turning on and off voices when relevant. In particular, it will host the core mechanism for turning off voices after every instrument and voice effect are finished.

The VoiceManager class is not finished yet, so it is not exposed in the ANGLECORE.h header.

Also added a renderer folder to prepare for the fourth agent of ANGLECORE: the Renderer.