Accelerated the rendering sequence calculation
The algorithm that calculates the rendering sequence of a Workflow now avoids exploring workers that are already part of the rendering sequence. An additional check is performed right at the beginning of each call to completeRenderingSequenceForWorker().

The ID-based comparison of a worker to the current worker has also been centralized into a lambda function, which is used twice in the completeRenderingSequenceForWorker() method, as the latter now relies more on the find_if method from the <algorithm> header to search for items.