Corrected the ParameterGenerator class and ANGLECORE header
After a first set of tests, 3 bugs were fixed in this commit:

1. Parameter change requests with duration 0 were not processed properly, as the generator would enter the wrong state. The generator now enters the state TRANSIENT_TO_STEADY (instead of STEADY) when receiving a request for an instantaneous transient, just as if we had just completed one.

2. It appeared that the parameter's value was not updated after each segment of the transient sequence was rendered, hereby causing the transient values to start again on each rendering session (although the generator would still complete its transient and reach its target value, as it counts the samples to track the transient phase).

3. Another issue was related to multiplicative transients starting on 0: in this case, we now artificially change the parameter's current value to a number slightly above zero in order for the geometric sequence to start and render properly.

Note that, because of precision issues, the ADDITIVE and MULTIPLICATIVE transients do not reach their target values at the same time even if they are assigned the same duration. Indeed, because the computation is virtually exact for the ADDITIVE method, the corresponding sequence reaches its target value one sample earlier than would its MULTIPLICATIVE counterpart (with the same configuration). This difference will not be audible at all, and should not impact the quality of the rendering.