Reimplemented the Instrument class
The Builder class no longer exists, so instruments now only inherits from the Worker class. This is because the creation of instruments and their insertion into an AudioWorkflow now follow a conventional, standardized process, which still leaves room for advanced customization within the Instrument core routine.

Indeed, instruments now rely on two main components that define their input interface (the output of an Instrument always consists of exactly as many streams as the number of channels):
- context parameters: these are the parameters that are generated away from the Instrument, by the GlobalContext or the VoiceContext, and which are shared with other instruments or effects. Examples include the sample rate, the reciprocal of the sample rate, and the frequency some instruments should play.
- parameters: these parameters, which are simply called parameters without any adjective, are the parameters that are specific to the Instrument, and used by the latter exclusively. This type of parameter requires to build a dedicated set of parameter generators, which each produce an input stream for the Instrument. Examples include, for an oscillator, the waveform, the number of unison voices, and the detune amount.

Also, an error was corrected in the VoiceContext which did not properly computed the ratio: frequency / sample rate.