Implemented and tested the Master's rendering routine and "note on" feature
ANGLECORE is now capable of both inserting a new instrument while rendering live, and playing that instrument when a note is pressed. The next step will be to implement the "note off" feature.

The entire code at this stage has been tested with a real audio application, and the tests were successful.

Note that for the test, a trivial version of the "note off" feature was implemented: the voices playing the note are all turned off instantly and abruptly, without cleaning out their left-over samples within the AudioWorkflow, which can cause glitches when the number of samples queried by the host increases as it will suddenly read the leftovers (this can be triggered by reducing the tempo, for instance).

Some work still needs to be done in therms of performance, such as:
- indicating to the Exporter the number of voices currently on, so it can stop its work when no voice is on and simply fill the output with zeros
- indicating to the Mixer which instrument rack is empty so that it can skip it (using the same system with increments as the existing one)
- removing unnecessary parameter generators, for instance for the sample rate and its reciprocal
- only compute some context parameters, such as the frequency over the sample rate, when new information is received, which requires to be able to tell when a generator has just generated a new series of values.

This is a major commit, as it incorporates many modifications, and proposes a first operational version of the Master class.