Implemented and tested the Master's "note off" feature
ANGLECORE is now capable of dealing with instruments that have a release stage in which they generate their audio tails. When a note is released, all the voices that are both on and playing the note are required to stop, but are not turned off immediately. Instead, the Master computes and keeps track of the number of samples the voices need to fade out, based on the length of each of their instrument's audio tail. When all of the instruments within a voice have completed their release stage, the voice is turned off.

Note that, to save computation time and ensure the instruments generate silence after their audio tail, a specific set of states was defined for the Instrument class. This set of states (ON, ON_ASKED_TO_STOP, ON_TO_OFF, OFF) automatically detects when an instrument is done with its audio tail, and turns it off, so it does not compute anything anymore and let the other instruments within the same voice finish their own audio tail. This modification now requires developers to implement a play() function in every subclasses of the Instrument class, instead of the previously pure virtual work() function.

The entire code at this stage has been tested with a real audio application, and the tests were successful.