Corrected the note velocity feature
Now instruments can adapt to the velocity they receive when a note is pressed.

The velocity is a discrete parameter, so there is no need to repeat its value all across a voice's velocity stream. Therefore, to save computation time, the velocity is only written at the first position of that stream when a key is pressed. The rest of the stream remains untouched.

This implies that instruments using velocity to generate their audio should only use the first sample of their velocity input stream, and ignore the other samples.