Implemented and tested the parameter feature
ANGLECORE is now capable of processing real-time changes of instrument parameters. When the end-user requests a new value for a parameter, the Master automatically sends that request to be taken care of in the next rendering session. During the latter, the request will be be processed, and the parameter's value will change accordingly. ANGLECORE will automatically smooth out any change of a parameter if the parameter has the "minimal smoothing" feature enabled.

In addition, since instruments from the same rack but on a different voice still need to share the same parameter values, parameter generators were grouped per rack and are no longer instrument-specific. Otherwise, if each instrument had its own set of parameter generators per voice, then a parameter change request may be processed as many times as they are voices, each voice processing it independently.

This commit introduces the concept of parameter registers and ParameterRegistrationPlan. A ParameterRegister is a register that keeps track of which ParameterGenerator corresponds to a given Parameter. It is consulted by the real-time thread when the Master needs to post a change request to the proper parameter generator. Since the register must be kept up to date each time the user either adds or removes an Instrument, it is also written into by the real-time thread on those occasions. A ParameterRegistrationPlan, which is precisely an update plan for a ParameterRegister, is used in these situations.

The entire code at this stage has been tested with a real audio application, and the tests were successful.

Some work still needs to be made. For instance, the Master should return immediately after a ParameterChangeRequest is processed, so that the future GUI components can update instantly, instead of waiting for the request-checking process to finish (which is when the Master copies the request, sends the copy of that request, and waits until the copy is destroyed so that it can destroy the original request on the non real-time thread, and avoid real-time memory deallocation).