Fixed crash when adding an instrument while a note is pressed
If an Instrument is added to the AudioWorkflow while a note is being pressed, then the instrument will automatically start and play. But the instruction to start requires all connections to be successfully done beforehand, which was not the case. This fix now performs all connections first, and instrument start-up second.

A similar fix will need to be implemented later for removing instruments, when the feature will be added. For the moment, removed instruments are just brutally disconnected from the AudioWorkflow. No crash should occur since none of the instrument's methods is called to stop the sound: the instrument will simply be excluded from the rendering sequence. But the sound cut will be harsh and a small tail off would be preferable.