Implemented a RequestManager class to handle requests asynchronously
The RequestManager class adds a new way of passing requests to the real-time thread, in addition to the simple, direct transaction previously implemented, which is still available. It now enables to pass connection and instrument requests to an intermediate, non real-time thread that pushes them into a waiting line before sending them to the real-time thread. This intermediate step provides additional safety, as the intermediate thread ensures requests are handled one at a time by the real-time thread. The RequestManager class also opens door to a new way for the Master to retrieve information on requests asynchronously, which will allow request-posting methods to return instantly, without blocking the user interface thread.