Fixed crash when adding instruments live
The AudioWorkflow needs to check if a rack is on before calling its instrument inside. However, the corresponding check was only performed on the instrument's existence, which actually does not guarantee its readiness to be used: the instrument may have not been connected to the rendering pipeline yet, which could cause a crash when called, as observed in practice during testing.

Therefore, a new boolean has been added to properly test whether a rack is on or not before calling the instrument inside.

Note that this boolean will soon be renamed, as the "on/off" terminology for racks will be replaced with a new "activated/deactivated" terminology.