Replaced the "on/off" terminology with "activated/deactivated" for racks
Racks are now referred to as "activated" or "deactivated" instead of being "on" or "off". This new terminology makes it clearer and more distinctive from instruments and voices, for which "on" and "off" status both implies a readiness to play on properly constructed racks, whereas "activated" and "deactivated" marks a clear differentiation between racks that are ready to be used, and those who are not.