Removed the Result sub-structure from the InstrumentRequest structure
Now the Master only reports a rack number after a new Instrument is added. The rack number indicates the rack where the Master will try to insert the new Instrument. If no slots were found to store a new instrument, then it will return the upper-bound index of all voices' racks (ANGLECORE_MAX_NUM_INSTRUMENTS_PER_VOICE), which corresponds to the maximum number of instruments that can be inserted in one AudioWorkflow, and which constitutes an out-of-range index easily detectable by the caller.